Allow fuller testing of AoF

+++ rmtest/disposableredis/__init__.py
@@ -3,9 +3,8 @@ import socket
 import redis
 import time
 import os
+import os.path
 import sys
-import itertools
-from contextlib import contextmanager
 
 
 REDIS_DEBUGGER = os.environ.get('REDIS_DEBUGGER', None)
@@ -48,6 +47,8 @@ class DisposableRedis(object):
         # It's equal to `_port` unless `_port` is None
         # in that case `port` is randomly generated
         self.port = None
+        self._is_external = True if port else False
+        self.use_aof = extra_args.pop('use_aof', False)
         self.extra_args = []
         for k, v in extra_args.items():
             self.extra_args.append('--%s' % k)
@@ -57,25 +58,31 @@ class DisposableRedis(object):
                 self.extra_args.append(v)
 
         self.path = path
-        self.dumped = False
         self.errored = False
+        self.dumpfile = None
+        self.aoffile = None
+        self.pollfile = None
 
     def _get_output(self):
         return '' if REDIS_SHOW_OUTPUT else self.process.stdout.read()
 
     def _start_process(self):
-        #print("Starting redis process: {}".format(' '.join(self.args)))
+        if self._is_external:
+            return
+
         if REDIS_DEBUGGER:
             debugger = REDIS_DEBUGGER.split()
             args = debugger + self.args
         else:
             args = self.args
         stdout = None if REDIS_SHOW_OUTPUT else subprocess.PIPE
+        if REDIS_SHOW_OUTPUT:
+            sys.stderr.write("Executing: {}".format(repr(args)))
         self.process = subprocess.Popen(
             args,
             stdin=sys.stdin,
             stdout=stdout,
-            stderr=sys.stderr
+            stderr=sys.stderr,
         )
 
         while True:
@@ -101,21 +108,35 @@ class DisposableRedis(object):
         else:
             self.port = self._port
 
-        self.dumpfile = 'dump.%s.rdb' % self.port
+        if not self.dumpfile:
+            self.dumpfile = 'dump.%s.rdb' % self.port
+        if not self.aoffile:
+            self.aoffile = 'appendonly.%s.aof' % self.port
+
         self.args = [self.path,
                      '--port', str(self.port),
                      '--save', '',
-                     '--dbfilename', self.dumpfile] + self.extra_args
+                     '--dbfilename', self.dumpfile]
+        if self.use_aof:
+            self.args += ['--appendonly', 'yes', '--appendfilename', self.aoffile]
 
+        self.args += self.extra_args
         self._start_process()
 
-    def stop(self):
-        self.process.terminate()
-        if self.dumped:
+    def _cleanup_files(self):
+        for f in (self.aoffile, self.dumpfile):
             try:
-                os.unlink(self.dumpfile)
+                os.unlink(f)
             except OSError:
                 pass
+
+    def stop(self, for_restart=False):
+        if self._is_external:
+            return
+
+        self.process.terminate()
+        if not for_restart:
+            self._cleanup_files()
 
     def __enter__(self):
         self.start()
@@ -126,18 +147,38 @@ class DisposableRedis(object):
         if exc_val or self.errored:
             sys.stderr.write("Redis output: {}\n".format(self._get_output()))
 
-    def dump_and_reload(self):
+    def _wait_for_child(self):
+        # Wait until file is available
+        r = self.client()
+        while True:
+            info = r.info('persistence')
+            if info['aof_rewrite_scheduled'] or info['aof_rewrite_in_progress']:
+                time.sleep(0.1)
+            else:
+                break
+
+    def dump_and_reload(self, restart_process=False):
         """
         Dump the rdb and reload it, to test for serialization errors
         """
         conn = self.client()
-        conn.save()
-        self.dumped = True
-        try:
-            conn.execute_command('DEBUG', 'RELOAD')
-        except redis.RedisError as err:
-            self.errored = True
-            raise err
+
+        if restart_process:
+            if self._is_external:
+                raise Exception('Tied to an external process - cannot restart')
+            import time
+            conn.bgrewriteaof()
+            self._wait_for_child()
+
+            self.stop(for_restart=True)
+            self.start()
+        else:
+            conn.save()
+            try:
+                conn.execute_command('DEBUG', 'RELOAD')
+            except redis.RedisError as err:
+                self.errored = True
+                raise err
 
     def client(self):
         """
